travelogue/pyproject.toml
[project]
name = "travelogue"
version = "0.1.0"
description = "journal for travelogue"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.7",
    "ollama>=0.6.1",
    "pillow>=12.0.0",
    "qdrant-client>=1.16.0",
    "streamlit>=1.51.0",
    "torch>=2.9.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
    "pyyaml>=6.0.0",
    "insightface>=0.7.3",
    "opencv-python>=4.11.0.86",
    "numpy>=2.3.4",
    "onnxruntime>=1.23.2",
    "bitsandbytes>=0.48.2",
    "accelerate>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

# If you also build sdists and want data included:
[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    # ... other includes like README, pyproject.toml, etc.
]

[tool.hatch.build.targets.wheel]
packages = ["src/travelogue"]
